fix: use uint64 for bitmask types to support 32-bit platforms
ngxHTTPOIDCConf (0x100000000) exceeds the range of uint on 32-bit
architectures where uint is 32 bits. Change all bitmask types from
uint to uint64 to ensure consistent behavior across platforms.

BREAKING CHANGE: MatchFunc signature changed from []uint to []uint64.

+++ analyze_lua_directives.gen.go
@@ -15,7 +15,7 @@
 
 package crossplane
 
-var luaDirectives = map[string][]uint{
+var luaDirectives = map[string][]uint64{
     "access_by_lua": {
         ngxHTTPMainConf | ngxHTTPSrvConf | ngxHTTPLocConf | ngxHTTPLifConf | ngxConfTake1,
     },
@@ -268,7 +268,7 @@ var luaDirectives = map[string][]uint{
 }
 
 // MatchLuaLatest is a MatchFunc for latest version of Lua.
-func MatchLuaLatest(directive string) ([]uint, bool) {
+func MatchLuaLatest(directive string) ([]uint64, bool) {
     m, ok := luaDirectives[directive]
     return m, ok
 }
